refactor: moved out upload results
